+++ tinted-builder/src/scheme/color.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 use std::fmt;
 
-use anyhow::{Context, Result};
+use anyhow::{anyhow, Context, Result};
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct Color {
@@ -12,28 +12,26 @@ pub struct Color {
 
 impl Color {
     pub fn new(hex_color: String) -> Result<Color> {
-        let hex_full = match process_hex_input(&hex_color) {
-            Some(valid_hex) => valid_hex,
-            None => {
-                anyhow::bail!("Provided hex value is not formatted correctly");
-            }
-        };
-
+        let hex_full = process_hex_input(&hex_color)
+            .ok_or_else(|| anyhow::anyhow!("Provided hex value is not formatted correctly"))?;
         let hex: (String, String, String) = (
             hex_full[0..2].to_lowercase(),
             hex_full[2..4].to_lowercase(),
             hex_full[4..6].to_lowercase(),
         );
         let rgb = hex_to_rgb(&hex)
-            .unwrap_or_else(|_| panic!("Unable to convert hex value to rgb: {}", hex_full));
-        let (r, g, b) = rgb;
-        let dec: (f32, f32, f32) = (r as f32 / 255.0, g as f32 / 255.0, b as f32 / 255.0);
+            .map_err(|_| anyhow!("Unable to convert hex value to rgb: {}", hex_full))?;
+        let dec: (f32, f32, f32) = (
+            rgb.0 as f32 / 255.0,
+            rgb.1 as f32 / 255.0,
+            rgb.2 as f32 / 255.0,
+        );
 
         Ok(Color { hex, rgb, dec })
     }
 
     pub fn to_hex(&self) -> String {
-        format!("#{}{}{}", &self.hex.0, &self.hex.1, &self.hex.2)
+        format!("{}{}{}", &self.hex.0, &self.hex.1, &self.hex.2)
     }
 }